requirements.txt
streamlit
pandas
plotly
google-auth
google-auth-oauthlib
google-api-python-client
python-dateutil
